docker/jupyterhub/requirements-jupyterhub.txt
# JupyterHub Requirements
# JupyterHub 4.0.2 and related packages

# JupyterHub core
jupyterhub==4.0.2
jupyterlab==4.0.11
notebook==7.0.8

# Kubernetes integration
jupyterhub-kubespawner==6.2.0
kubernetes==28.1.0

# Authentication
jupyterhub-ldapauthenticator==1.3.2
jupyterhub-nativeauthenticator==1.2.0
oauthenticator==16.2.1

# Database support
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# HTTP and networking
tornado==6.4
aiohttp==3.9.1
requests==2.31.0

# Configuration and templating
jinja2==3.1.2
traitlets==5.14.0
configurable-http-proxy==4.5.4

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Data processing (for admin tasks)
pandas==2.1.4
numpy==1.26.2

# Web UI enhancements
jupyterhub-announcement==1.1.0

# Development and debugging
ipython==8.18.1
ipdb==0.13.13

# Async utilities
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Additional spawner options
dockerspawner==12.1.0
systemdspawner==1.0.1

# REST API enhancements
jupyterhub-idle-culler==1.2.1

# SSL/TLS support
pyopenssl==23.3.0

# Time zone handling
pytz==2023.4